Consolidate inputs and drop AutoDataset
