+++ flash/core/data/data_module.py
@@ -23,12 +23,10 @@ from torch.utils.data.dataset import IterableDataset, Subset
 from torch.utils.data.sampler import Sampler
 
 import flash
-from flash.core.data.auto_dataset import BaseAutoDataset, IterableAutoDataset
 from flash.core.data.base_viz import BaseVisualization
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_pipeline import DataPipeline
-from flash.core.data.io.input import Input
-from flash.core.data.io.input_base import InputBase, IterableInput
+from flash.core.data.io.input import Input, InputBase, IterableInput
 from flash.core.data.io.input_transform import DefaultInputTransform, InputTransform
 from flash.core.data.io.output_transform import OutputTransform
 from flash.core.data.splits import SplitDataset
@@ -272,7 +270,7 @@ class DataModule(pl.LightningDataModule):
             self.set_dataset_attribute(self._predict_ds, "running_stage", RunningStage.PREDICTING)
 
     def _resolve_collate_fn(self, dataset: Dataset, running_stage: RunningStage) -> Optional[Callable]:
-        if isinstance(dataset, (BaseAutoDataset, SplitDataset, InputBase)):
+        if isinstance(dataset, (SplitDataset, InputBase)):
             return self.data_pipeline.worker_input_transform_processor(running_stage)
 
     def _train_dataloader(self) -> DataLoader:
@@ -280,7 +278,7 @@ class DataModule(pl.LightningDataModule):
         train_ds: Dataset = self._train_ds() if isinstance(self._train_ds, Callable) else self._train_ds
         shuffle: bool = False
         collate_fn = self._resolve_collate_fn(train_ds, RunningStage.TRAINING)
-        if isinstance(train_ds, (IterableAutoDataset, IterableInput)):
+        if isinstance(train_ds, IterableInput):
             drop_last = False
         else:
             drop_last = len(train_ds) > self.batch_size
@@ -289,7 +287,7 @@ class DataModule(pl.LightningDataModule):
 
         if self.sampler is None:
             sampler = None
-            shuffle = not isinstance(train_ds, (IterableDataset, IterableAutoDataset, IterableInput))
+            shuffle = not isinstance(train_ds, (IterableDataset, IterableInput))
         else:
             sampler = self.sampler(train_ds)
 
@@ -374,7 +372,7 @@ class DataModule(pl.LightningDataModule):
         """Configure the prediction dataloader of the datamodule."""
         predict_ds: Dataset = self._predict_ds() if isinstance(self._predict_ds, Callable) else self._predict_ds
 
-        if isinstance(predict_ds, (IterableAutoDataset, IterableInput)):
+        if isinstance(predict_ds, IterableInput):
             batch_size = self.batch_size
         else:
             batch_size = min(self.batch_size, len(predict_ds) if len(predict_ds) > 0 else 1)
@@ -468,7 +466,7 @@ class DataModule(pl.LightningDataModule):
         if not isinstance(val_split, float) or (isinstance(val_split, float) and val_split > 1 or val_split < 0):
             raise MisconfigurationException(f"`val_split` should be a float between 0 and 1. Found {val_split}.")
 
-        if isinstance(train_dataset, (IterableAutoDataset, IterableInput)):
+        if isinstance(train_dataset, IterableInput):
             raise MisconfigurationException(
                 "`val_split` should be `None` when the dataset is built with an IterableDataset."
             )
