+++ flash/image/classification/adapters.py
@@ -28,9 +28,7 @@ from torch.utils.data import DataLoader, IterableDataset, Sampler
 
 import flash
 from flash.core.adapter import Adapter, AdapterTask
-from flash.core.data.auto_dataset import BaseAutoDataset
-from flash.core.data.io.input import DataKeys
-from flash.core.data.io.input_base import InputBase
+from flash.core.data.io.input import DataKeys, InputBase
 from flash.core.model import Task
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.compatibility import accelerator_connector
@@ -197,7 +195,7 @@ class Learn2LearnAdapter(Adapter):
     def _convert_dataset(
         self,
         trainer: "flash.Trainer",
-        dataset: BaseAutoDataset,
+        dataset: InputBase,
         ways: int,
         shots: int,
         queries: int,
@@ -205,7 +203,7 @@ class Learn2LearnAdapter(Adapter):
         num_task: int,
         epoch_length: int,
     ):
-        if isinstance(dataset, (InputBase, BaseAutoDataset)):
+        if isinstance(dataset, InputBase):
 
             metadata = getattr(dataset, "data", None)
             if metadata is None or (metadata is not None and not isinstance(dataset.data, list)):
@@ -325,7 +323,7 @@ class Learn2LearnAdapter(Adapter):
 
     def process_train_dataset(
         self,
-        dataset: BaseAutoDataset,
+        dataset: InputBase,
         trainer: "flash.Trainer",
         batch_size: int,
         num_workers: int,
@@ -362,7 +360,7 @@ class Learn2LearnAdapter(Adapter):
 
     def process_val_dataset(
         self,
-        dataset: BaseAutoDataset,
+        dataset: InputBase,
         trainer: "flash.Trainer",
         batch_size: int,
         num_workers: int,
@@ -399,7 +397,7 @@ class Learn2LearnAdapter(Adapter):
 
     def process_test_dataset(
         self,
-        dataset: BaseAutoDataset,
+        dataset: InputBase,
         trainer: "flash.Trainer",
         batch_size: int,
         num_workers: int,
@@ -436,7 +434,7 @@ class Learn2LearnAdapter(Adapter):
 
     def process_predict_dataset(
         self,
-        dataset: BaseAutoDataset,
+        dataset: InputBase,
         batch_size: int = 1,
         num_workers: int = 0,
         pin_memory: bool = False,
