+++ flash/image/classification/integrations/baal/data.py
@@ -19,8 +19,8 @@ import torch
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from torch.utils.data import DataLoader, Dataset, random_split
 
-from flash.core.data.auto_dataset import BaseAutoDataset
 from flash.core.data.data_pipeline import DataPipeline
+from flash.core.data.io.input import InputBase
 from flash.core.data.new_data_module import DataModule
 from flash.core.utilities.imports import _BAAL_AVAILABLE, requires
 
@@ -36,11 +36,11 @@ else:
         pass
 
 
-def dataset_to_non_labelled_tensor(dataset: BaseAutoDataset) -> torch.tensor:
+def dataset_to_non_labelled_tensor(dataset: InputBase) -> torch.tensor:
     return np.zeros(len(dataset))
 
 
-def filter_unlabelled_data(dataset: BaseAutoDataset) -> Dataset:
+def filter_unlabelled_data(dataset: InputBase) -> Dataset:
     return dataset
